+++ BE/src/diaries/diaries.controller.ts
@@ -14,12 +14,12 @@ import {
   CreateDiaryDto,
   DeleteDiaryDto,
   DiaryUuidDto,
-  ReadDiaryDto,
   UpdateDiaryDto,
-} from "./diaries.dto";
+} from "./dto/diaries.dto";
 import { Diary } from "./diaries.entity";
 import { AuthGuard } from "@nestjs/passport";
 import { IdGuard } from "src/auth/auth.id-guard";
+import { ReadDiaryDto } from "./dto/diaries.read.dto";
 
 @Controller("diaries")
 @UseGuards(AuthGuard())
